pyproject.toml
[tool.poetry]
name = "hyms-plat"
version = "0.1.0"
description = "Portal de Hinários do Santo Daime"
authors = ["Nitai Bezerra"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
wagtail = "^6.4"
psycopg2-binary = "^2.9"
typesense = "^0.21"
django-allauth = "^65.3"
celery = "^5.4"
redis = "^5.2"
weasyprint = "^63"
pillow = "^11.1"
python-magic = "^0.4"
pyyaml = "^6.0"
gunicorn = "^23.0"
whitenoise = "^6.8"
django-environ = "^0.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
pytest-cov = "^6.0"
factory-boy = "^3.3"
playwright = "^1.49"
black = "^24.10"
isort = "^5.13"
ruff = "^0.8"
pre-commit = "^4.0"
django-debug-toolbar = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip = ["migrations", ".venv"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["migrations", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
fail_under = 80
